+++ pix-editor/app/routes/authenticated/static-courses/new.js
@@ -1,3 +1,13 @@
 import Route from '@ember/routing/route';
+import { inject as service } from '@ember/service';
 
-export default class StaticCourseNewRoute extends Route {}
+export default class StaticCourseNewRoute extends Route {
+  @service access;
+  @service router;
+
+  beforeModel() {
+    if (!this.access.mayCreateOrEditStaticCourse()) {
+      this.router.transitionTo('authenticated.static-courses.list');
+    }
+  }
+}
